+++ src/fields/EntriesSubset.php
@@ -16,9 +16,12 @@ use nfourtythree\entriessubset\services\EntriesSubsetService;
 
 use Craft;
 use craft\base\ElementInterface;
+use craft\db\Table;
 use craft\elements\Entry;
 use craft\fields\BaseRelationField;
+use craft\helpers\Db;
 use craft\helpers\Json;
+use craft\fields\Entries;
 
 /**
  *  Field
@@ -33,7 +36,7 @@ use craft\helpers\Json;
  * @package   EntriesSubset
  * @since     1.0.0
  */
-class EntriesSubsetField extends BaseRelationField
+class EntriesSubsetField extends Entries
 {
     // Public Properties
     // =========================================================================
@@ -125,7 +128,9 @@ class EntriesSubsetField extends BaseRelationField
         $entryTypes = $settings[ 'entryTypes' ];
 
         if ($entryTypes and is_array( $entryTypes ) and !empty( $entryTypes ) ) {
-          foreach( $entryTypes as $typeId ) {
+          foreach( $entryTypes as $typeUid ) {
+            $typeId = Db::idByUid(Table::ENTRYTYPES, $typeUid);
+
             if ( is_numeric( $typeId ) ) {
               $entryType = Craft::$app->sections->getEntryTypeById( $typeId );