+++ src/services/EntriesSubsetService.php
@@ -54,7 +54,7 @@ class EntriesSubsetService extends Component
       if ( !isset( $entryTypeOptions[ $type->section->handle ] ) ) {
         $entryTypeOptions[ $type->section->handle ] = [];
       }
-      $entryTypeOptions[ $type->section->handle ][] = [ 'label' => $type->name, 'value' => $type->id ];
+      $entryTypeOptions[ $type->section->handle ][] = [ 'label' => $type->name, 'value' => $type->uid ];
     }
 
     return $entryTypeOptions;
@@ -96,6 +96,7 @@ class EntriesSubsetService extends Component
 
     return $users;
   }
+
 
   /**
    * Get Categories