+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.2.2 - 2019-09-06
+### Added
+- Added Project Config support for Entry Types
+
+## 1.2.1 - 2019-04-18
+### Fixed
+- Updated field to extend Entries class fixes #9 h/t @tinchalamo
+- Added support for Reverse Related Entries plugin
+
+## 1.2.0 - 2019-01-31
+### Fixed
+- Fixed support for Craft CMS 3.1
+
 ## 1.1.0 - 2018-04-27
 ### Added
 - Added ability to restrict entry selection by entry Author